requirements.txt
agent-framework-github-copilot
agent-framework-claude
python-dotenv
